add doubles to stat command + fix stat issues

+++ src/main/java/com/skyblock/skyblock/utilities/Util.java
@@ -353,7 +353,7 @@ public class Util {
         return stack;
     }
 
-    public String abbreviate(int num) {
+    public String abbreviate(double num) {
         if (num < 1000) return num + "";
         int exp = (int) (Math.log(num) / Math.log(1000));
         return String.format("%.1f%c", num / Math.pow(1000, exp), "kMGTPE".charAt(exp - 1)).replaceAll("\\.0", "");
